requirements.txt
Flask
fastapi
python-jose
passlib
bcrypt
pymongo
python-multipart
